Enable short-circuiting eval for And/Or
short circuiting was turned off, under the assumption that Diff and Translate change the order of asserted conjuncts (refer issue #53). However, the real issue was with Viper's implementation of short-circuiting eval rule which used to swap conjuncts in the case of && (before it was corrected in Silicon). Corrected implementation removed swapIfAnd (source of issue).